# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "lessonlink-backend バックエンドAPI",
        "title": "lessonlink-backend",
        "contact": {},
        "license": {
            "name": "typedef-tokyo."
        },
        "version": "1.0"
    },
    "host": "localhost:3002",
    "basePath": "/",
    "paths": {
...
        "/schedule/list/{campus}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "スケジュールリスト取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "校舎",
                        "name": "campus",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.ScheduleListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/schedule/{schedule_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "スケジュールリスト取得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ScheduleID",
                        "name": "schedule_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "履歴番号",
                        "name": "history",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.ScheduleGetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "スケジュール保存",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ScheduleID",
                        "name": "schedule_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "スケジュール保存リクエスト",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ScheduleSaveRequestData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.ScheduleSaveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "スケジュール削除",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ScheduleID",
                        "name": "schedule_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/schedule/{schedule_id}/item-return-list": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "スケジュール編集アイテムリスト移動",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ScheduleID",
                        "name": "schedule_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "アイテムリスト移動リクエスト",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ScheduleItemReturnListRequestData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.ScheduleItemEditResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/schedule/{schedule_id}/item-shift": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "スケジュール編集アイテムシフト",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ScheduleID",
                        "name": "schedule_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "アイテムシフトリクエスト",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ScheduleItemShiftRequestData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.ScheduleItemEditResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.ScheduleItemReturnListRequestData": {
            "type": "object",
            "required": [
                "duration",
                "history_index",
                "identifier",
                "lesson_id"
            ],
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "history_index": {
                    "type": "integer"
                },
                "identifier": {
                    "type": "string"
                },
                "lesson_id": {
                    "type": "integer"
                }
            }
        },
        "controller.ScheduleItemShiftRequestData": {
            "type": "object",
            "required": [
                "history_index",
                "room_index"
            ],
            "properties": {
                "history_index": {
                    "type": "integer"
                },
                "room_index": {
                    "type": "integer"
                }
            }
        },
        "controller.ScheduleSaveRequestData": {
            "type": "object",
            "required": [
                "history_index"
            ],
            "properties": {
                "history_index": {
                    "type": "integer"
                }
            }
        },
...
        "presenter.ScheduleGetResponse": {
            "type": "object",
            "required": [
                "campus",
                "created_user_id",
                "history_index",
                "lesson_item_list",
                "room_lesson_list",
                "rooms",
                "schedule_end_time",
                "schedule_id",
                "schedule_start_time",
                "title"
            ],
            "properties": {
                "campus": {
                    "type": "string"
                },
                "created_user_id": {
                    "type": "integer"
                },
                "history_index": {
                    "type": "integer"
                },
                "lesson_item_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.ScheduleLessonItem"
                    }
                },
                "room_lesson_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.ScheduleRoomLesson"
                    }
                },
                "rooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.ScheduleRoomDTO"
                    }
                },
                "schedule_end_time": {
                    "type": "integer"
                },
                "schedule_id": {
                    "type": "integer"
                },
                "schedule_start_time": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "presenter.ScheduleItemEditLessonItem": {
            "type": "object",
            "required": [
                "duration",
                "identifier",
                "lesson_id",
                "lesson_name"
            ],
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "identifier": {
                    "type": "string"
                },
                "lesson_id": {
                    "type": "integer"
                },
                "lesson_name": {
                    "type": "string"
                }
            }
        },
        "presenter.ScheduleItemEditResponse": {
            "type": "object",
            "required": [
                "history_index",
                "lesson_item_list",
                "room_lesson_list"
            ],
            "properties": {
                "history_index": {
                    "type": "integer"
                },
                "lesson_item_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.ScheduleItemEditLessonItem"
                    }
                },
                "room_lesson_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.ScheduleItemEditRoomLesson"
                    }
                }
            }
        },
        "presenter.ScheduleItemEditRoomLesson": {
            "type": "object",
            "required": [
                "duration",
                "end_time_hour",
                "end_time_minutes",
                "identifier",
                "item_tag",
                "lesson_id",
                "lesson_name",
                "room_index",
                "start_time_hour",
                "start_time_minutes"
            ],
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "end_time_hour": {
                    "type": "integer"
                },
                "end_time_minutes": {
                    "type": "integer"
                },
                "identifier": {
                    "type": "string"
                },
                "item_tag": {
                    "type": "string"
                },
                "lesson_id": {
                    "type": "integer"
                },
                "lesson_name": {
                    "type": "string"
                },
                "room_index": {
                    "type": "integer"
                },
                "start_time_hour": {
                    "type": "integer"
                },
                "start_time_minutes": {
                    "type": "integer"
                }
            }
        },
        "presenter.ScheduleLessonItem": {
            "type": "object",
            "required": [
                "duration",
                "identifier",
                "lesson_id",
                "lesson_name"
            ],
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "identifier": {
                    "type": "string"
                },
                "lesson_id": {
                    "type": "integer"
                },
                "lesson_name": {
                    "type": "string"
                }
            }
        },
        "presenter.ScheduleListDTO": {
            "type": "object",
            "required": [
                "created_user_id",
                "created_user_name",
                "last_update_date_time",
                "last_update_user_name",
                "schedule_id",
                "title"
            ],
            "properties": {
                "created_user_id": {
                    "type": "integer"
                },
                "created_user_name": {
                    "type": "string"
                },
                "last_update_date_time": {
                    "type": "string"
                },
                "last_update_user_name": {
                    "type": "string"
                },
                "schedule_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "presenter.ScheduleListResponse": {
            "type": "object",
            "required": [
                "schedules"
            ],
            "properties": {
                "schedules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.ScheduleListDTO"
                    }
                }
            }
        },
        "presenter.ScheduleRoomDTO": {
            "type": "object",
            "required": [
                "room_index",
                "room_name",
                "visible"
            ],
            "properties": {
                "room_index": {
                    "type": "integer"
                },
                "room_name": {
                    "type": "string"
                },
                "visible": {
                    "type": "boolean"
                }
            }
        },
        "presenter.ScheduleRoomLesson": {
            "type": "object",
            "required": [
                "duration",
                "end_time_hour",
                "end_time_minutes",
                "identifier",
                "item_tag",
                "lesson_id",
                "lesson_name",
                "room_index",
                "start_time_hour",
                "start_time_minutes"
            ],
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "end_time_hour": {
                    "type": "integer"
                },
                "end_time_minutes": {
                    "type": "integer"
                },
                "identifier": {
                    "type": "string"
                },
                "item_tag": {
                    "type": "string"
                },
                "lesson_id": {
                    "type": "integer"
                },
                "lesson_name": {
                    "type": "string"
                },
                "room_index": {
                    "type": "integer"
                },
                "start_time_hour": {
                    "type": "integer"
                },
                "start_time_minutes": {
                    "type": "integer"
                }
            }
        },
        "presenter.ScheduleSaveResponse": {
            "type": "object",
            "required": [
                "history_index",
                "msg"
            ],
            "properties": {
                "history_index": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}